[#10] refactor: make dummy data code refactor
